Added more sound effects

+++ UnityFantasyRpg/Assets/Scripts/Enemy Script/EnemyHealthBar.cs
@@ -8,11 +8,11 @@ public class EnemyHealthBar : MonoBehaviour
     public Slider slider;
     public GameObject enemy;
 
-    public void SetMaxHealth(int health){
+    public void SetMaxHealth(float health){
         slider.maxValue = health;
         slider.value = health;
     }
-    public void SetHealth(int health){
+    public void SetHealth(float health){
         slider.value = health;
     }
